requirements.txt
streamlit==1.35.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
